+++ src/keycodes/aliases.js
@@ -1,0 +1,21 @@
+export default {
+  windows: 91,
+  '⇧': 16,
+  '⌥': 18,
+  '⌃': 17,
+  '⌘': 91,
+  ctl: 17,
+  control: 17,
+  option: 18,
+  pause: 19,
+  break: 19,
+  caps: 20,
+  return: 13,
+  escape: 27,
+  spc: 32,
+  pgup: 33,
+  pgdn: 34,
+  ins: 45,
+  del: 46,
+  cmd: 91
+}
