+++ src/keycodes/numpad.js
@@ -1,0 +1,20 @@
+export default {
+  'numpad *': 106,
+  // 'numpad +': 107,
+  'numpad +': 43,
+  'numpad add': 43, // as a trick
+  'numpad -': 109,
+  'numpad .': 110,
+  'numpad /': 111,
+  'num lock': 144,
+  'numpad 0': 96,
+  'numpad 1': 97,
+  'numpad 2': 98,
+  'numpad 3': 99,
+  'numpad 4': 100,
+  'numpad 5': 101,
+  'numpad 6': 102,
+  'numpad 7': 103,
+  'numpad 8': 104,
+  'numpad 9': 105
+}
